- mkhive's "registry.c" used wcsncmp for comparing two strings. But this function seems to work different on different hosts (probably due to different handling of wchar_t's or whatever).   To avoid this problem, use memcmp now to compare the strings as it's done two lines later. This fixes the Live-CD on Linux/Unix hosts.   (For the first time, Live-CDs created by our 64-Bit Debug Buildslave are usable :-) - Change __volatile to volatile in "winddk.h". This fixes compilation problems under MSVC. - Make the prototype of RtlCompareMemory in "winnt.h" similar to the one in "winddk.h". Fixes a compilation problem in MSVC. - Define the prototype for xwcschr in "mkhive.h". Silences a warning in MSVC. - Add some casts to "reginf.c" to silence some warnings in MSVC. - Add a MSVC 2005 project file for mkhive. mkhive can be compiled under MSVC with this project file, when cmlib and inflib_host are compiled. - Remove the duplicate definition of CMLIB_HOST in "cmlib.mak".
svn path=/trunk/; revision=28425

+++ reactos/tools/mkhive/reginf.c
@@ -262,7 +262,7 @@ do_reg_operation(HKEY KeyHandle,
 	      if (Str == NULL)
 		return FALSE;
 
-	      InfHostGetMultiSzField (Context, 5, Str, Size, NULL);
+	      InfHostGetMultiSzField (Context, 5, Str, (ULONG)Size, NULL);
 	    }
 
 	  if (Flags & FLG_ADDREG_APPEND)
@@ -291,7 +291,7 @@ do_reg_operation(HKEY KeyHandle,
 	      if (Str == NULL)
 		return FALSE;
 
-	      InfHostGetStringField (Context, 5, Str, Size, NULL);
+	      InfHostGetStringField (Context, 5, Str, (ULONG)Size, NULL);
 	    }
 	}
 
@@ -347,7 +347,7 @@ do_reg_operation(HKEY KeyHandle,
 	    return FALSE;
 
 	  DPRINT("setting binary data %s len %lu\n", ValueName, Size);
-	  InfHostGetBinaryField (Context, 5, Data, Size, NULL);
+	  InfHostGetBinaryField (Context, 5, Data, (ULONG)Size, NULL);
 	}
 
       RegSetValueExA (KeyHandle,
@@ -396,7 +396,7 @@ registry_callback (HINF hInf, PCHAR Section, BOOL Delete)
 
       /* get key */
       Length = strlen (Buffer);
-      if (InfHostGetStringField (Context, 2, Buffer + Length, MAX_INF_STRING_LENGTH - Length, NULL) != 0)
+      if (InfHostGetStringField (Context, 2, Buffer + Length, MAX_INF_STRING_LENGTH - (ULONG)Length, NULL) != 0)
 	*Buffer = 0;
 
       DPRINT("KeyName: <%s>\n", Buffer);
